+++ core/include/core/SCNN.h
@@ -15,16 +15,16 @@ namespace core {
     protected:
 
         /* Number of PE columns */
-        const int Wt;
+        const uint32_t Wt;
 
         /* Number of PE rows */
-        const int Ht;
+        const uint32_t Ht;
 
         /* Column multipliers per PE */
-        const int I;
+        const uint32_t I;
 
         /* Row multipliers per PE */
-        const int F;
+        const uint32_t F;
 
         /* Output accumulator size */
         const int out_acc_size;
@@ -38,7 +38,7 @@ namespace core {
          * @param y         W position
          * @return          Accumulator bank index
          */
-        int map_accumulator(int k, int x, int y);
+        int map_accumulator(uint32_t k, uint32_t x, uint32_t y);
 
     private:
 
@@ -69,7 +69,7 @@ namespace core {
          * @param wgt       1D weights queue with linearized activations indexes to be processed
          * @return          Return stats for the given PE
          */
-        PE_stats computeSCNNPE(int W, int H, int stride, const act_idxMap &act, const wgt_idxMap &wgt);
+        PE_stats computeSCNNPE(uint64_t W, uint64_t H, int stride, const act_idxMap &act, const wgt_idxMap &wgt);
 
         /* Compute SCNN tile
          * @param n         Number of batch
@@ -92,9 +92,9 @@ namespace core {
          * @param wgt       Weights for the layer
          * @param stats     Statistics to fill
          */
-        void computeSCNNTile(int n, int ct, int ck, int kc, int tw, int th, int X, int Y, int Kc, int K, int W, int H,
-                int R, int S, int stride, int padding, const cnpy::Array<T> &act, const cnpy::Array<T> &wgt,
-                sys::Statistics::Stats &stats);
+        void computeSCNNTile(int n, int ct, int ck, int kc, int tw, int th, uint64_t X, uint64_t Y, int Kc, uint64_t K,
+                uint64_t W, uint64_t H, uint64_t R, uint64_t S, int stride, int padding, const cnpy::Array<T> &act,
+                const cnpy::Array<T> &wgt, sys::Statistics::Stats &stats);
 
         /* Compute the timing for a layer
          * @param layer     Layer for which we want to calculate the outputs
@@ -130,8 +130,8 @@ namespace core {
          * @param _N_THREADS    Number of parallel threads for multi-threading execution
          * @param _FAST_MODE    Enable fast mode to simulate only one image
          */
-        SCNN(int _Wt, int _Ht, int _I, int _F, int _out_acc_size, int _BANKS, uint8_t _N_THREADS,
-            bool _FAST_MODE) : Simulator<T>(_N_THREADS, _FAST_MODE), Wt(_Wt), Ht(_Ht), I(_I), F(_F),
+        SCNN(uint32_t _Wt, uint32_t _Ht, uint32_t _I, uint32_t _F, uint32_t _out_acc_size, uint32_t _BANKS,
+            uint8_t _N_THREADS, bool _FAST_MODE) : Simulator<T>(_N_THREADS, _FAST_MODE), Wt(_Wt), Ht(_Ht), I(_I), F(_F),
             out_acc_size(_out_acc_size), BANKS(_BANKS) {}
 
         /* Run the timing simulator of the architecture
